+++ requirements.txt
@@ -6,7 +6,7 @@ Werkzeug==0.11.10
 
 git+https://github.com/AusDTO/dto-digitalmarketplace-utils.git@25.14.0#egg=dto-digitalmarketplace-utils==25.14.0
 git+https://github.com/AusDTO/dto-digitalmarketplace-content-loader.git@1.5.1#egg=dto-digitalmarketplace-content-loader==1.5.1
-git+https://github.com/AusDTO/dto-digitalmarketplace-apiclient.git@6.7.0#egg=dto-digitalmarketplace-apiclient==6.7.0
+git+https://github.com/AusDTO/dto-digitalmarketplace-apiclient.git@6.8.0#egg=dto-digitalmarketplace-apiclient==6.8.0
 
 newrelic==2.68.0.50
 cffi==1.5.2
